+++ services/collectd/config.go
@@ -27,10 +27,18 @@ const (
 
 	DefaultTypesDB = "/usr/share/collectd/types.db"
 
-	// DefaultUDPReadBuffer is the default UDP read buffer
+	// DefaultReadBuffer is the default buffer size for the UDP listener.
 	// Sets the size of the operating system's receive buffer associated with
-	// the UDP traffic
-	DefaultReadBuffer = 8 * 1024 * 1024
+	// the UDP traffic. Keep in mind that the OS must be able
+	// to handle the number set here or the UDP listener will error and exit.
+	//
+	// DefaultReadBuffer = 0 means to use the OS default, which is usually too
+	// small for high UDP performance.
+	//
+	// Increasing OS buffer limits:
+	//     Linux:      sudo sysctl -w net.core.rmem_max=<read-buffer>
+	//     BSD/Darwin: sudo sysctl -w kern.ipc.maxsockbuf=<read-buffer>
+	DefaultReadBuffer = 0
 )
 
 // Config represents a configuration for the collectd service.